icons replacement in inventory and hotbar
allows to use seperate png files to display items in hotbar and
inventory.  a few blocks do not look great when in inventory or hotbar,
this makes for a good alternative.

only issue is that on mouse_click and drag it shows the normal textures.
fix or help with this if you can.