Add automatic checks of the pep8 using flake8 and Github Actions
Like #35, I believe that project is a great educational tool, an impressive achievement, and a reference for clean Python code.

Warnings F403 and F405 are currently ignored. That could be removed when #124 is merged.

Warnings E128,E201,E226,E231 could be removed if autopep8 or similar is used. I advise to merge #124 before, to limit the risk of merge conflict.

I fixed 3 small pep8 issues in main.py too.